refactor: Move resources related things to the resources file
Visibility changes to `pub`:
- all fields in `Resources`
- the `load_resources` function
